worked on deleting a product image when product is deleted

+++ controllers/admin.ts
@@ -3,6 +3,7 @@
 import { Response, Request, NextFunction } from "express";
 import { validationResult } from "express-validator/check";
 import { Product } from "../models";
+import { deleteFile } from "../util/file";
 // import { Model } from "sequelize-typescript";
 
 // import Product from "../models/product";
@@ -200,6 +201,7 @@ export const postEditProduct = (
         product.price = price;
         product.description = description;
         if (image) {
+          deleteFile(product.imageUrl);
           product.imageUrl = image.path;
         }
         return product.save().then(() => {
@@ -233,7 +235,14 @@ export const postDeleteProduct = (
   next: NextFunction
 ) => {
   const prodId = req.body?.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product): any => {
+      if (!product) {
+        return next(new Error("Product not found"));
+      }
+      deleteFile(product?.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("Product deleted");
       res.redirect("/admin/products");
